% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.offshore.R
\name{pptv.offshore.live}
\alias{pptv.offshore.live}
\title{Generate numbers given in the report for Live, Full Episode}
\usage{
pptv.offshore.live(pptv_data, View = FALSE)
}
\arguments{
\item{pptv_data}{PPTV data set}

\item{View}{(optional; {TRUE|FALSE} or {0|1})
By default (FALSE) data is returned to parent environment
(whether that is to a variable or to screen). By
flagging View (TRUE), the data is instead routed to 
View()}
}
\value{
Live, Full Episode metrics and KPIs for
the Raw and Smackdown episodes aired most recently
relative to the user-provided as_on_dt.
}
\description{
This function returns the Live unique viewer and minute counts which correspond
to the most recently aired Raw and Smackdown full-length episodes 
relative to the provided as_on_dt. Additionally, it returns the
"Vs 4-Week Avg" KPIs found in report generated by the offshore
EXL team:
}
\details{
COMPUTE TRAILING 4-WEEK AVERAGE
trlAvg = avg(cnt_1wkAgo, cnt_2wksAgo, cnt_3wksAgo, cnt_4wksAgo)
  
COMPUTE THE DELTA (PERCENT DIFF BETWEEN THIS WEEK AND TRLNG AVG )
thisWksCnt - trlAvg            
--------------------- * 100\%   
trlAvg

The script was designed (in conjunction with most of the pptvR package)
to reverse engineer the numbers given in the PPTV report generated by the 
offshore EXL team.  For as long as the offshore EXL team is in charge
of generating the report, this script serves as a double
check to ensure no copy-and-paste errors have made their 
way into the report, which has happened from time to time.
}
\examples{
data = pptv.report.live("2017-03-09")
         
}
