% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.offshore.R
\name{pptv.offshore.overview}
\alias{pptv.offshore.overview}
\title{Compute numbers for "Total Viewership" and "Live Viewership" presented
on last page of Offshore Report.}
\usage{
pptv.offshore.overview(pptv_data, type = "Total", all_vars = FALSE)
}
\arguments{
\item{pptv_data}{Must input a PPTV data set, e.g., pptv_data = pptv.excel(as_on_dt).}

\item{type}{(default: "Total")
Specify either "Total Viewership" or "Live Viewership".  Note that
this parameter is not case sensitive and only depends on the first
letter (e.g., "LIVE" and "l" will both yield numbers for "Live 
Viewership").}

\item{all_vars}{Default=FALSE. If TRUE, returns uniques too.}
}
\description{
This function will give you the numbers for Raw, Smackdown, Other,
and Total.
}
