% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.offshore.R
\name{pptv.offshore}
\alias{pptv.offshore}
\title{Return table of values as would be seen in Offshore's PPTV report.}
\usage{
pptv.offshore(pptv_data)
}
\arguments{
\item{pptv_data}{}
}
\description{
Bla, bla
}
