% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.offshore.R
\name{pptv.offshore.shortform}
\alias{pptv.offshore.shortform}
\title{Return table of values as would be seen on Short Form page of
Offshore's PPTV report.}
\usage{
pptv.offshore.shortform(pptv_data, N = 3, v2 = TRUE, all_vars = FALSE)
}
\arguments{
\item{pptv_data}{}

\item{N}{}

\item{v2}{In mid-May 2017, we changed the format of this page quite a bit.
Set to FALSE to view old way of doing it.}

\item{all_vars}{}
}
\description{
Bla, bla
}
