% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.offshore.R
\name{pptv.offshore.topmatches}
\alias{pptv.offshore.topmatches}
\title{Top N Cuts}
\usage{
pptv.offshore.topmatches(pptv_data, N = 20, to_csv = FALSE)
}
\arguments{
\item{pptv_data}{}

\item{N}{}

\item{to_csv}{}
}
\description{
Bla, bla
}
