% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.offshore.R
\name{pptv.offshore.web}
\alias{pptv.offshore.web}
\title{Compute aggreagted PPTV website numbers.}
\usage{
pptv.offshore.web(as_on_dt = Sys.Date(), day1 = "Monday", v2 = TRUE)
}
\arguments{
\item{as_on_dt}{}

\item{day1}{}

\item{ExlWrongWay}{}
}
\description{
Compute aggreagted PPTV website numbers.
}
