% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.qc.R
\name{pptv.qc.clipVsFull}
\alias{pptv.qc.clipVsFull}
\title{Analyze show titles in PPTV data sets.}
\usage{
pptv.qc.clipVsFull(pptvData)
}
\arguments{
\item{pptvData}{This function only takes pptvData.}
}
\value{
Show titles...
}
\description{
This function figures out if show titles are consistent between
content versions full (version=='Full Show') and clip (version!='Full Show').
}
