% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.qc.R
\name{pptv.qc.ljNowThen}
\alias{pptv.qc.ljNowThen}
\title{Check whether the metrics and KPIs corresponding each row in
the inner join between as_on_dt1 and as_on_dt2 are consistent.}
\usage{
pptv.qc.ljNowThen(pptv_data1, pptv_data2, nulls = FALSE, pcDiffs = FALSE,
  View = FALSE)
}
\arguments{
\item{pptv_data1}{PPTV data set}

\item{pptv_data2}{PPTV data set}

\item{nulls}{(default: FALSE)
Defaults to joining on all PPTV dimensional parameters. However,
sometimes PPTV does not specify usAirDate ("MatchDate"), and so
usAirDate shows up as NULL.  By setting nulls=TRUE, we...}

\item{pcDiffs}{...}

\item{View}{(optional; {TRUE|FALSE} or {0|1})
By default (FALSE) data is returned to parent environment
(whether that is to a variable or to screen). By
flagging View (TRUE), the data is instead routed to 
View()}
}
\description{
This function does not restrict the data to any subsets, nor does
this function compare the consistency of aggregated values. 
This is a brute force row-to-row comparison, where the compared
rows are those found by taking the inner join on PPTV viewership
dimensions (chinaAirDate, dataDate, )
We have found that inconsistencies may exist in the untouched 
data that do not affect all or any aggregated values of interest,
but are still disconcerting and require discovery just in case.
For example, in the 2017-03-09 data set, we found that...
}
