% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.qc.R
\name{pptv.qc.lp7}
\alias{pptv.qc.lp7}
\title{Check whether the LIVE+7 metrics and KPIs in the inner join between
as_on_dt1 and as_on_dt2 are the same, or if they have changed.}
\usage{
pptv.qc.lp7(pptv_data1, pptv_data2, trail = 0, View = FALSE)
}
\arguments{
\item{pptv_data1}{PPTV data set}

\item{pptv_data2}{PPTV data set}

\item{trail}{(optional; INTEGER)
By default, all live data is  retrieved.  Restrict 
to trailing N weeks by specifying integer N.  Note
that any decimal are automatically floored.}

\item{View}{(optional; {TRUE|FALSE} or {0|1})
By default (FALSE) data is returned to parent environment
(whether that is to a variable or to screen). By
flagging View (TRUE), the data is instead routed to 
View()}

\item{more}{(optional)}
}
\description{
This function restricts its attention to just Raw and 
Smackdown because, ultimately, it is a check on metrics and KPIs
included in the report generated by the offshore EXL team.  
For this reason, it also restricted to LIVE+7 measures
with the hope that such modularization might help us more quickly 
identify the "what" of "when something is wrong."
}
\details{
ATTRIBUTES OF LIVE+7 VIEWERSHIP:
  (a) show=={'Mon Night Raw'|'Smackdown'}
  
  (b) version=='Full Show'
      -- clips are analyzed elsewhere in the PPTV 
         report
  
  (c) audio=={'Mandarin'|'English'}
      -- audio=='NULL' in the raw data is a placeholder
         for version=='Others', which do not currently
         distinguish which language was viewed; i.e.,
         version=='Full Show' should have no audio=='NULL'
         by definition in the data guide provided by the
         PPTV team, so we can safely discard
  
  (d) dataDate >= chinaAirDate+days(6)
      -- we only want to consider rows that have data
         recorded in a 7-day interval: the LIVE date
         for type == {'LIVE' | 'VOD'} (i.e., 
         dataDate==chinaAirDate) and the 6 days thereafter 
         (chinaAirData + seq(6), which on has type='VOD')
  
  (*) subtitles: No Restrictions (i.e., 'NULL' or 'QF')
      -- for audio=='Mandarin', subtitles=='NULL', however
         for audio=='English', subtitles=={'NULL'|'QF'},
         which we will refer to as eng0 and eng1 for 
         "English w/ no subtitles" and "English w/ subtitles"
  
  (*) type: No Restrisctions ('LIVE' or 'VOD')
      -- LIVE is always has dataDate == chinaAirDate ==usAirDate+1,
         however internally to PPTV it must literally reference
         the exact datetime the show was LIVE b/c there also exists
         VOD viewership on the LIVE date
      -- VOD has any dataDate 
      -- so, technically, we want count both together
}
