% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.qc.R
\name{pptv.qc.nulls}
\alias{pptv.qc.nulls}
\title{Checks for nulls in the metrics/KPIs in a PPTV data set.}
\usage{
pptv.qc.nulls(pptv_data, View = FALSE)
}
\arguments{
\item{pptv_data}{PPTV data set}

\item{View}{(optional; {TRUE|FALSE} or {0|1})
By default (FALSE) data is returned to parent environment
(whether that is to a variable or to screen). By
flagging View (TRUE), the data is instead routed to 
View()}
}
\description{
The metrics/KPIs are:  
    totView,  totUniq,  totMin, vpu,  mpv,  mpu
}
\details{
NOTE: pptv.nullcheck only looks for NULLs in the metrics/KPIs.
This is because sometimes a dimension parameter, like usAirDate,
is assigned a NULL value (e.g., clips that did not air in 
the U.S. do not have finite usAirDate values).
 
We found that data with finite values in the 2017-01-19 
data set no showed up with NULL values in the 2017-01-26
data set.  In other data sets, we have found extraneous 
rows with all NULL values.
}
