% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.qc.R
\name{pptv.qc}
\alias{pptv.qc}
\title{Run PPTV Quality Control (QC) Checks}
\usage{
pptv.qc(pptv_data1, pptv_data2)
}
\arguments{
\item{pptv_data1}{}

\item{pptv_data2}{}
}
\description{
This function will print notes to screen as its runs, and
ultimately will return a list of the QC checks.
}
