% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.query.R
\name{pptv.query}
\alias{pptv.query}
\title{Specify as_on_dt to load most recent PPTV daily data.
 
EXCEL:  For the first several months, we received PPTV data
in Excel files only.  This was the original data
importation function used to bring PPTV data into R.}
\usage{
pptv.query(as_on_dt, con = NA, web = FALSE, sdstar = FALSE)
}
\arguments{
\item{as_on_dt}{(character or Date)
User may provide any date in the form YYYY-MM-DD, 
however it will be converted to the nearest, 
previous as_on_dt for which PPTV provided a data 
set.}

\item{sdstar}{If TRUE, then "Smackdown" and "Smackdown*" are distinguishable 
show titles.  By default, "Smackdown*" is converted to "Smackdown".}
}
\value{
All PPTV data with given as_on_dt.
}
\description{
REDSHIFT:  At one point, we got things into RedShift,
however there were several hiccups.  The early stuff
might not be trustworthy (unless DE went back and
fixed those).  However, anything after say 2017-04-xx
is probably good!
}
\details{
NOTE:  pptv.query.excel, pptv.redshift, and other helper functions
  are available to view and edit in the pptv.query text file.
}
\examples{
data = pptv.query.excel("2017-03-09")
         
}
