% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.query.R
\name{pptv.query.web}
\alias{pptv.query.web}
\title{Specify as_on_dt to load most recent PPTV data set.
 
For the first several months, we received PPTV data
in Excel files only.  This was the original data
importation function used to bring PPTV data into R.}
\usage{
pptv.query.web(as_on_dt = Sys.Date(), agg = FALSE, day1 = "Monday")
}
\arguments{
\item{as_on_dt}{Defaults to Sys.Date(), which is converted to nearest
valid as_on_dt previous to Sys.Date().
User may provide any date in the form YYYY-MM-DD, 
which will be converted to the nearest valid
as_on_dt, if necessary.}
}
\value{
All PPTV web data with given as_on_dt (or nearest valid 
as_on_dt previous to provided as_on_dt).
}
\description{
Specify as_on_dt to load most recent PPTV data set.
 
For the first several months, we received PPTV data
in Excel files only.  This was the original data
importation function used to bring PPTV data into R.
}
\examples{
data = pptv.webdata("2017-03-09")
         
}
