% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.query.R
\name{pptv.showTitle}
\alias{pptv.showTitle}
\title{Clean up Chinese characters in Show Titles.}
\usage{
pptv.showTitle(pptv_data)
}
\arguments{
\item{pptv_data}{This function only takes pptv_data.}

\item{sdstar}{Defaults to FALSE. If true, the SD* shows are separated
from the SD shows.  For some reason within the first month
or so of PPTV data, this was a thing.  I'm pretty sure you won't
have to ever set this keyword to TRUE, but it's there just in
case.}
}
\value{
pptv_data w/ formatted showTitles data.
}
\description{
This function ensures readability of the Show Titles.
}
