% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.viz.live.R
\name{pptv.viz.live}
\alias{pptv.viz.live}
\title{LIVE for Raw and Smackdown}
\usage{
pptv.viz.live(pptv_data, width = 9, height = 6, legend = FALSE,
  to_png = "", to_pdf = "")
}
\arguments{
\item{pptv_data}{}

\item{width}{Defaults to 9 inches (fits PowerPoint template well)}

\item{height}{Defaults to 6 inches (fits PowerPoint template well)}

\item{to_png}{}

\item{to_pdf}{Looks the best!}
}
\description{
LIVE for Raw and Smackdown
}
