% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.viz.R
\name{pptv.viz}
\alias{pptv.viz}
\title{Multiple plot function}
\usage{
pptv.viz(..., plotlist = NULL, file, cols = 1, layout = NULL,
  width = 4.8, height = 4.8, to_png = "", to_pdf = "", newWay = TRUE)
}
\arguments{
\item{...}{Pass in N ggplot objects as the first N positional parameters.}

\item{plotlist}{(default: NULL)
Pass in a list of ggplot objects using this named parameter.
Note: You can use this indpendently of or in conjunction with
the first N positional parameters.}

\item{file}{Not Yet Used in Code. In Future: Option to save figure automatically.}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.}

\item{width}{Defaults to 480.}

\item{height}{Defaults to 480.}

\item{to_png}{}
}
\description{
Plot multiple ggplot objects in the same figure.  Pass as manu
ggplot objects in as individual inputs (...), or use the plotlist
keyboard to input them in a list.
}
