% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.viz.short.R
\name{pptv.viz.short}
\alias{pptv.viz.short}
\title{Total shortform viewership for Raw, Smackdown, and Total}
\usage{
pptv.viz.short(pptv_data, width = 9, height = 5, to_png = "",
  to_pdf = "")
}
\arguments{
\item{pptv_data}{}

\item{width}{Defaults to 9 inches (fits PowerPoint template well)}

\item{height}{Defaults to 6 inches (fits PowerPoint template well)}

\item{to_png}{}

\item{to_pdf}{Looks the best!}
}
\description{
Total shortform viewership for Raw, Smackdown, and Total
}
