% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.viz.ubound.R
\name{pptv.viz.ubound}
\alias{pptv.viz.ubound}
\title{For LIVE+7, generate ribbon plot that bounds the uniques.}
\usage{
pptv.viz.ubound(pptv_data, t0 = as.Date("2016-06-27"),
  tf = as.Date(Sys.time()), yscale = 5e+05, ymax = NULL, rmv = 64,
  smv = 46, width = 9, height = 6, to_png = "", to_pdf = "",
  v2 = TRUE)
}
\arguments{
\item{pptv_data}{}

\item{ymax}{}

\item{rmv}{}

\item{smv}{}

\item{width}{}

\item{height}{}
}
\description{
Since a "unique" in the PPTV is actually only unique up
to the particular asset and day (e.g., show, showTitle, 
audio, subtitle, device, day), we cannot truly get an 
estimate of total unique viewers.  However, we can bound
that estimate.
}
\details{
The count of total uniques itself represents
and overshoot --- an upper bound.  Depending on how many
people decide to watch a particular episode of Raw on
more than one device, on more than one day, in multiple
languages, etc, this can be an extremely bad upper bound.
However, for a given day and asset, the total views are never 
too much more than the total uniques, so it seems likely 
that this is not too bad of an upper bound.

For a lower bound, we used Nielsen data to get an estimate
of how long people typically watch Raw or Smackdown.  Specifically,
we were provided with the average viewing duration for each episode
over the lifetime of PPTV.  We took the median of these averages for
each show:  64 minutes for Raw and 46 minutes for Smackdown.
By default, the lower bound in this figure is the total number of
minutes divided by this number, giving a curve that represents
the number of viewers given all viewers watched for the typical
duration.
}
