% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptv.viz.webcal.R
\name{pptv.viz.webcal}
\alias{pptv.viz.webcal}
\title{Plot data from Webpage data file on a calendar.}
\usage{
pptv.viz.webcal(as_on_dt, N = 3, device = "Web", type = "Unique")
}
\arguments{
\item{as_on_dt}{....}

\item{N}{(default: 3)
Specify the number N of most recent months to display.}

\item{device}{(default: "Web")
Specify the device of interest: "Web", "App", or 
"Total".  Note that this parameter is not case
sensitive and dependent only on first 3 characters
of device name.}

\item{type}{(default: unique)
Specify the type of data: "Page Views" or "Unique Views".
Note that this parameter is not case sensitive and is
only dependent on the first character (e.g., type="u"
and type="UNIQUE" both specify "Unique Views").}
}
\description{
Look at PPTV webpage activity in the familiar setting
of a calendar.  The calendar shows one device ("Web", 
"App", or "Total") and one metric ("UV" or "PV" for 
unique views or page views, respectively).  The default
is to display the most recent 3 calendar months, but 
this can be tweaked with the N parameter.
}
